feat: :sparkles: :recycle: abstraction to zenodo workflow
